added news spam approval

+++ module/scraper_notices.py
@@ -4,10 +4,14 @@ import ast
 import os
 import copy
 import yaml
+import time
+
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 
 with open('config/settings.yaml', 'r') as yaml_config:
     config_map = yaml.load(yaml_config)
     notices_urls = config_map["notices_urls"]
+    approve_group_chatid = config_map["approve_group_chatid"]
 
 def get_links(label, url):
     req = requests.get(url)
@@ -63,7 +67,6 @@ def get_notice_content(notice_dict, base_url, archive_p, notice_p):
     with open(notice_p, 'w') as fw:
         fw.write(formatted_notice)
 
-
 def spam_news(bot, notice_p, channel):
     if os.path.isfile(notice_p):
         message = open(notice_p).read()
@@ -72,6 +75,44 @@ def spam_news(bot, notice_p, channel):
                 bot.sendMessage(chat_id=channel, text=message, parse_mode='HTML')
             except Exception as error:
                 open("logs/errors.txt", "a+").write("{} {}\n".format(error, channel))
+        os.remove(notice_p)
+
+def spam_news_direct(bot, notice_message, channel):
+    if notice_message != "":
+        try:
+            bot.sendMessage(chat_id=channel, text=notice_message, parse_mode='HTML')
+        except Exception as error:
+            open("logs/errors.txt", "a+").write("{} {}\n".format(error, channel))
+
+def send_news_approve_message(bot, notice_p, channel_folder, pending_approval_folder, channel, group_chatid):
+    if os.path.isfile(notice_p):
+        notice_message = open(notice_p).read()
+
+        if notice_message != "":
+            try:
+                notice_disk_id = time.clock()
+                approving_notice_filename = "{}/{}/{}_{}.dat".format(channel_folder, pending_approval_folder, channel, notice_disk_id)
+
+                if not os.path.exists(os.path.dirname(approving_notice_filename)):
+                    try:
+                        os.makedirs(os.path.dirname(approving_notice_filename))
+                    except OSError as exc: 
+                        if exc.errno != errno.EEXIST:
+                            raise
+
+                with open(approving_notice_filename, 'w') as fw:
+                    fw.writelines(notice_message[0:])
+
+                keyboard_markup = [
+                    [InlineKeyboardButton("Accetta", callback_data="news:approved:{}:{}:{}".format(channel, channel_folder, notice_disk_id)),
+                    InlineKeyboardButton("Rifiuta", callback_data="news:rejected:{}:{}:{}".format(channel, channel_folder, notice_disk_id))]
+                ]
+
+                reply_markup = InlineKeyboardMarkup(keyboard_markup)
+
+                bot.sendMessage(chat_id=group_chatid, text=notice_message, parse_mode='HTML', reply_markup=reply_markup)
+            except Exception as error:
+                open("logs/errors.txt", "a+").write("send_news_approve_message: {} {}\n".format(error, channel))
         os.remove(notice_p)
 
 
@@ -87,8 +128,8 @@ def scrape_notices(bot, job):
         else:
             folder = i
 
-        pending_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avvisi.dat"
-        archive_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avvisi_in_sospeso.dat"
+        pending_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avvisi_in_sospeso.dat"
+        archive_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avvisi.dat"
         notice_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avviso.dat"
 
         base_url = notices_urls_cp[i]["urls"][list(notices_urls_cp[i]["urls"])[0]]
@@ -123,4 +164,5 @@ def scrape_notices(bot, job):
             if pending_notice:
                 get_notice_content(pending_notice, base_url, archive_path, notice_path)
         
-        spam_news(bot, notice_path, notices_urls[i]["channel"])
+        send_news_approve_message(bot, notice_path, "data/avvisi/"+str(folder), "in_approvazione", notices_urls[i]["channel"], approve_group_chatid)
+        # spam_news(bot, notice_path, notices_urls[i]["channel"])
